requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
django-rest-swagger==2.2.0
django-swagger-ui==0.1.11
djangorestframework==3.11.1
gunicorn==20.0.4
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
psycopg2==2.8.6
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.7
whitenoise==5.2.0
